Create ingestion utils module
The ingestion specific util functions in the main utils module are now
in this new ingestion specific utils module. Additionally the hash
calculating code has been moved into this new module.
